# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dealer-settings:
    get:
      operationId: getDealerSettings
      tags: [dealer-settings]
      summary: Get the public dealer profile
      description: Returns the configured dealer identity, contact details, showroom copy, services, and legal links.
      responses:
        "200":
          description: Dealer settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealerSettings"
        "500":
          description: Unable to load dealer settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateDealerSettings
      tags: [dealer-settings]
      summary: Replace the dealer profile
      description: Saves the complete dealer profile used by the public showroom and dealer portal.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DealerSettingsInput"
      responses:
        "200":
          description: Updated dealer settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealerSettings"
        "400":
          description: Dealer settings failed validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "500":
          description: Unable to update dealer settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    DealerOnlineReservation:
      type: object
      additionalProperties: false
      description: Online reservations are disabled until the dealership explicitly enables them and supplies its reservation terms. Payment is currently simulated.
      properties:
        enabled: { type: boolean }
        depositPence: { type: number, multipleOf: 1, minimum: 100, maximum: 1000000 }
        terms: { type: string, maxLength: 4000 }
      required: [enabled, depositPence, terms]
    DealerBrandColors:
      type: object
      additionalProperties: false
      properties:
        primaryHsl: { type: string, minLength: 1 }
        accentHsl: { type: string, minLength: 1 }
      required: [primaryHsl, accentHsl]
    DealerIdentity:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 120 }
        logoText: { type: string, maxLength: 120 }
        logoAsset: { type: string, maxLength: 500 }
        brandColors:
          $ref: "#/components/schemas/DealerBrandColors"
      required: [name, logoText, logoAsset, brandColors]
    DealerContact:
      type: object
      additionalProperties: false
      properties:
        phone: { type: string, maxLength: 40 }
        whatsapp: { type: string, maxLength: 40 }
        email: { type: string, maxLength: 160 }
      required: [phone, whatsapp, email]
    DealerAddress:
      type: object
      additionalProperties: false
      properties:
        street: { type: string, maxLength: 160 }
        city: { type: string, maxLength: 80 }
        region: { type: string, maxLength: 80 }
        postcode: { type: string, maxLength: 20 }
        mapsUrl: { type: string, maxLength: 500 }
      required: [street, city, region, postcode, mapsUrl]
    DealerHoursItem:
      type: object
      additionalProperties: false
      properties:
        days: { type: string, minLength: 1, maxLength: 80 }
        times: { type: string, minLength: 1, maxLength: 80 }
      required: [days, times]
    DealerLegal:
      type: object
      additionalProperties: false
      properties:
        companyName: { type: string, maxLength: 160 }
        companyNumber: { type: string, maxLength: 80 }
        vatNumber: { type: string, maxLength: 80 }
        termsUrl: { type: string, maxLength: 500 }
        privacyUrl: { type: string, maxLength: 500 }
        cookieUrl: { type: string, maxLength: 500 }
      required: [companyName, companyNumber, vatNumber, termsUrl, privacyUrl, cookieUrl]
    DealerSocial:
      type: object
      additionalProperties: false
      properties:
        instagram: { type: string, maxLength: 500 }
        facebook: { type: string, maxLength: 500 }
        twitter: { type: string, maxLength: 500 }
      required: [instagram, facebook, twitter]
    DealerHero:
      type: object
      additionalProperties: false
      properties:
        announcement: { type: string, maxLength: 160 }
        copy: { type: string, minLength: 1, maxLength: 160 }
        subcopy: { type: string, minLength: 1, maxLength: 300 }
        primaryCta: { type: string, minLength: 1, maxLength: 80 }
        secondaryCta: { type: string, minLength: 1, maxLength: 80 }
      required: [announcement, copy, subcopy, primaryCta, secondaryCta]
    DealerService:
      type: object
      additionalProperties: false
      properties:
        enabled: { type: boolean }
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, minLength: 1, maxLength: 300 }
        ctaLabel: { type: string, minLength: 1, maxLength: 80 }
      required: [enabled, title, description, ctaLabel]
    DealerTestDriveHours:
      type: object
      additionalProperties: false
      properties:
        day: { type: number, multipleOf: 1, minimum: 0, maximum: 6 }
        enabled: { type: boolean }
        open: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
        close: { type: string, pattern: "^([01][0-9]|2[0-3]):[0-5][0-9]$" }
      required: [day, enabled, open, close]
    DealerTestDriveBooking:
      type: object
      additionalProperties: false
      properties:
        enabled: { type: boolean }
        durationMinutes: { type: number, multipleOf: 1, minimum: 15, maximum: 180 }
        bufferMinutes: { type: number, multipleOf: 1, minimum: 0, maximum: 120 }
        minimumNoticeHours: { type: number, minimum: 0, maximum: 720 }
        dailyCapacity: { type: number, multipleOf: 1, minimum: 1, maximum: 100 }
        daysAhead: { type: number, multipleOf: 1, minimum: 1, maximum: 90 }
        blockedDates:
          type: array
          maxItems: 366
          uniqueItems: true
          items: { type: string, pattern: "^[0-9]{4}-[0-9]{2}-[0-9]{2}$" }
        weeklyHours:
          type: array
          minItems: 7
          maxItems: 7
          items: { $ref: "#/components/schemas/DealerTestDriveHours" }
        instructions: { type: string, maxLength: 2000 }
        confirmationMode: { type: string, enum: [instant, approval] }
      required: [enabled, durationMinutes, bufferMinutes, minimumNoticeHours, dailyCapacity, daysAhead, blockedDates, weeklyHours, instructions, confirmationMode]
...
    DealerBookViewing:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 160 }
        description: { type: string, minLength: 1, maxLength: 300 }
        ctaLabel: { type: string, minLength: 1, maxLength: 80 }
      required: [title, description, ctaLabel]
    DealerWhyBuyItem:
      type: object
      additionalProperties: false
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, minLength: 1, maxLength: 300 }
      required: [title, description]
    DealerRecentHandovers:
      type: object
      additionalProperties: false
      properties:
        enabled:
          type: boolean
        count:
          type: number
          multipleOf: 1
          minimum: 1
          maximum: 6
      required: [enabled, count]
...
    CustomerReview:
      type: object
      additionalProperties: false
      required: [rating, review, name, date, source]
      properties:
        verified: { type: boolean, description: "Verification confirmed by the dealership; not a Google certification." }
        invited: { type: boolean, description: "Customer was invited to leave this review." }
        rating: { type: number, enum: [1, 2, 3, 4, 5] }
        review: { type: string, minLength: 1, maxLength: 5000 }
        name: { type: string, minLength: 1, maxLength: 150 }
        date: { type: string, maxLength: 100 }
        source: { type: string, maxLength: 100 }
    DealerWebsiteCopy:
      type: object
      additionalProperties: false
      description: Optional plain-text dealership page copy. Blank values use the standard wording.
      properties:
        navigationStock: { type: string, maxLength: 1000 }
        navigationContact: { type: string, maxLength: 1000 }
        navigationPartExchange: { type: string, maxLength: 1000 }
        stockTitle: { type: string, maxLength: 1000 }
        stockIntroduction: { type: string, maxLength: 1000 }
        servicesEyebrow: { type: string, maxLength: 1000 }
        servicesHeading: { type: string, maxLength: 1000 }
        featuredHeading: { type: string, maxLength: 1000 }
        visitHeading: { type: string, maxLength: 1000 }
        teamHeading: { type: string, maxLength: 1000 }
        reviewsHeading: { type: string, maxLength: 1000 }
        contactTitle: { type: string, maxLength: 1000 }
        contactIntroduction: { type: string, maxLength: 1000 }
        contactTalkHeading: { type: string, maxLength: 1000 }
        contactVisitHeading: { type: string, maxLength: 1000 }
        contactVisitDescription: { type: string, maxLength: 1000 }
        contactDirectionsHeading: { type: string, maxLength: 1000 }
        contactBeforeHeading: { type: string, maxLength: 1000 }
        contactParkingHeading: { type: string, maxLength: 1000 }
        contactMessageHeading: { type: string, maxLength: 1000 }
        contactMessageDescription: { type: string, maxLength: 1000 }
        warrantyEyebrow: { type: string, maxLength: 1000 }
        warrantySupportingCopy: { type: string, maxLength: 1000 }
        warrantyDetailsHeading: { type: string, maxLength: 1000 }
        warrantyDetailsIntroduction: { type: string, maxLength: 1000 }
        warrantyEnquiryHeading: { type: string, maxLength: 1000 }
        savedEyebrow: { type: string, maxLength: 1000 }
        savedTitle: { type: string, maxLength: 1000 }
        savedDescription: { type: string, maxLength: 1000 }
        compareTitle: { type: string, maxLength: 1000 }
        compareDescription: { type: string, maxLength: 1000 }
        viewingTitle: { type: string, maxLength: 1000 }
        viewingDescription: { type: string, maxLength: 1000 }
        generalTitle: { type: string, maxLength: 1000 }
        generalDescription: { type: string, maxLength: 1000 }
        deliveryTitle: { type: string, maxLength: 1000 }
        deliveryDescription: { type: string, maxLength: 1000 }
        warrantyTitle: { type: string, maxLength: 1000 }
        warrantyDescription: { type: string, maxLength: 1000 }
        part_exchangeTitle: { type: string, maxLength: 1000 }
        part_exchangeDescription: { type: string, maxLength: 1000 }
        vehicleSpecificationHeading: { type: string, maxLength: 1000 }
        vehicleDescriptionHeading: { type: string, maxLength: 1000 }
        vehicleInformationHeading: { type: string, maxLength: 1000 }
        vehicleFeaturesHeading: { type: string, maxLength: 1000 }
        vehicleSimilarHeading: { type: string, maxLength: 1000 }
    DealerPresentation:
      type: object
      additionalProperties: false
      description: Optional showroom content. Empty values are not presented as factual claims.
      properties:
        pageColour: { type: string, pattern: "^(#[0-9a-fA-F]{6})?$" }
        panelColour: { type: string, pattern: "^(#[0-9a-fA-F]{6})?$" }
        headingColour: { type: string, pattern: "^(#[0-9a-fA-F]{6})?$" }
        linkColour: { type: string, pattern: "^(#[0-9a-fA-F]{6})?$" }
        featuredEnabled: { type: boolean }
        visitEnabled: { type: boolean }
        servicesEnabled: { type: boolean }
        showHeroDescription: { type: boolean }
        websiteCopy: { $ref: "#/components/schemas/DealerWebsiteCopy" }
        footerLogoUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        faviconUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        heroImagePosition: { type: string, enum: [center, left, right, top, bottom] }
        visitImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        visitImageAlt: { type: string, maxLength: 200 }
        contactImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        contactImageAlt: { type: string, maxLength: 200 }
        receptionImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        receptionImageAlt: { type: string, maxLength: 200 }
        reviewsEnabled: { type: boolean }
        reviews:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/CustomerReview"
        comparisonEnabled: { type: boolean, description: "Show customer vehicle comparison. Off when omitted." }
        heroImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        heroImageAlt: { type: string, maxLength: 200 }
        showroomImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        showroomImageAlt: { type: string, maxLength: 200 }
        teamImageUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        teamImageAlt: { type: string, maxLength: 200 }
        teamIntroduction: { type: string, maxLength: 1000 }
        visitInstructions: { type: string, maxLength: 1000 }
        parkingInstructions: { type: string, maxLength: 1000 }
        reviewsUrl: { type: string, maxLength: 2048, pattern: "^(https://[^\\s]+)?$" }
        includedInformation: { type: string, maxLength: 1000 }
    DealerBrochure:
      type: object
      additionalProperties: false
      description: Optional PDF presentation settings. Vehicle facts and insurance history are always included.
      properties:
        title: { type: string, maxLength: 60 }
        introduction: { type: string, maxLength: 300 }
        accentColour: { type: string, pattern: "^#[0-9a-fA-F]{6}$" }
        footerNote: { type: string, maxLength: 500 }
        includeDescription: { type: boolean }
        includeFeatures: { type: boolean }
        includeGallery: { type: boolean }
        photoLimit: { type: number, multipleOf: 1, minimum: 1, maximum: 80 }
        galleryLayout: { type: string, enum: [grid, large] }
    DealerSettings:
      type: object
      additionalProperties: false
      properties:
        presentation: { $ref: "#/components/schemas/DealerPresentation" }
        brochure: { $ref: "#/components/schemas/DealerBrochure" }
        identity: { $ref: "#/components/schemas/DealerIdentity" }
        contact: { $ref: "#/components/schemas/DealerContact" }
        address: { $ref: "#/components/schemas/DealerAddress" }
        hours:
          type: array
          maxItems: 14
          items: { $ref: "#/components/schemas/DealerHoursItem" }
        legal: { $ref: "#/components/schemas/DealerLegal" }
        social: { $ref: "#/components/schemas/DealerSocial" }
        hero: { $ref: "#/components/schemas/DealerHero" }
        featuredVehicleIds:
          type: array
          maxItems: 8
          uniqueItems: true
          items: { type: string, minLength: 1, maxLength: 120 }
        warranty: { $ref: "#/components/schemas/DealerService" }
        delivery: { $ref: "#/components/schemas/DealerService" }
        partExchange: { $ref: "#/components/schemas/DealerService" }
        onlineReservation: { $ref: "#/components/schemas/DealerOnlineReservation" }
        bookViewing: { $ref: "#/components/schemas/DealerBookViewing" }
        testDriveBooking: { $ref: "#/components/schemas/DealerTestDriveBooking" }
        recentHandovers: { $ref: "#/components/schemas/DealerRecentHandovers" }
        trustItems:
          type: array
          maxItems: 8
          items: { type: string, minLength: 1, maxLength: 120 }
        whyBuy:
          type: array
          maxItems: 8
          items: { $ref: "#/components/schemas/DealerWhyBuyItem" }
      required: [identity, contact, address, hours, legal, social, hero, featuredVehicleIds, warranty, delivery, partExchange, bookViewing, recentHandovers, trustItems, whyBuy]
    DealerSettingsInput:
      allOf:
        - $ref: "#/components/schemas/DealerSettings"
...
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error:
          type: string
          minLength: 1
      required:
        - error
...